+++ backend/src/main/java/com/project/Tamago/security/Token.java
@@ -1,4 +1,4 @@
-package com.project.Tamago.util.security;
+package com.project.Tamago.security;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
