+++ backend/src/main/java/com/project/Tamago/service/AuthService.java
@@ -1,7 +1,7 @@
 package com.project.Tamago.service;
 
 import static com.project.Tamago.exception.exceptionHandler.ErrorCode.*;
-import static com.project.Tamago.util.constants.Constant.*;
+import static com.project.Tamago.constants.Constant.*;
 
 import java.util.concurrent.TimeUnit;
 
@@ -19,9 +19,9 @@ import com.project.Tamago.dto.requestDto.JoinReqDto;
 import com.project.Tamago.dto.requestDto.LoginReqDto;
 import com.project.Tamago.dto.requestDto.PasswordReqDto;
 import com.project.Tamago.exception.CustomException;
-import com.project.Tamago.jwt.JwtTokenProvider;
+import com.project.Tamago.security.jwt.JwtTokenProvider;
 import com.project.Tamago.repository.UserRepository;
-import com.project.Tamago.util.security.Token;
+import com.project.Tamago.security.Token;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
